+++ freight/app_settings.py
@@ -21,3 +21,26 @@ FREIGHT_DISCORD_PING_TYPE = getattr(
     'FREIGHT_DISCORD_PING_TYPE', 
     None
 )
+
+# modes of operation for Alliance Freight
+FREIGHT_OPERATION_MODE_MY_ALLIANCE = 'my_alliance'
+FREIGHT_OPERATION_MODE_MY_CORPORATION = 'my_corporation'
+
+FREIGHT_OPERATION_MODES = [
+    (FREIGHT_OPERATION_MODE_MY_ALLIANCE, 'My Alliance'),
+    (FREIGHT_OPERATION_MODE_MY_CORPORATION, 'My Corporation'),        
+]
+
+if (hasattr(settings, 'FREIGHT_OPERATION_MODE') 
+    and settings.FREIGHT_OPERATION_MODE in [x[0] for x in FREIGHT_OPERATION_MODES]):
+    FREIGHT_OPERATION_MODE = settings.FREIGHT_OPERATION_MODE
+else:
+    FREIGHT_OPERATION_MODE = FREIGHT_OPERATION_MODE_MY_ALLIANCE
+
+def get_freight_operation_mode_friendly(mode: str) -> str:
+    """returns user friendly description of operation mode"""    
+    msg = [(x, y) for x, y in FREIGHT_OPERATION_MODES if x == mode]
+    if len(msg) != 1:
+        raise ValueError('Undefined mode')
+    else:
+        return msg[0][1]